Raycast detecting objects and display UI with Item Info
raycast is now connected with the Shop event of opening the UI and displaying the information of the item that the player is looking at

TO DO LIST
add a buy script: adds items to inventory
add a sell feature: need access to the inventory system
add a script that increases or decrease amount you want to buy or sell
get inventory system completed so that items can be added when purchased.

suggested ideas
detect when player isn't moving to stop raycast?
[could match the vector LO]
does game pause when menu pops up could allow player to click buttons.