update inventory system
inventory system is functional with adding items as stacks or as single items.

items can selected when clicked
items can be removed from inventory when clicked again.

TO DO List
have items be able to be used in inventory and do their effect be used in game.

inventory system is basically finished at this point and is ready to be merged into develop after merging develop into inventory feature.