introduction of the ShopKeeper
shopkeeper AI follows player in the shop

TO DO
need to have a pop up for sell System to open UI shop/inventory to sell items and increase players currency
have items be purchasable

DEV NOTES
wont have the items be able to be bought in bulk as time restraint.
hoping to have all this done by Friday.