inventory pauses game in shop
when opening inventory in the Shop scene it prevents the raycast to detect anything when you look around.
when inventory opens, your mouse becomes visible and invisible when in game.

DEV NOTE
when adding prefabs/items to stock in the shop system make sure you add "itemSelection" tag to the shop item as the ray cast checks for this.